refactor: Changed the description position of some functions and classes.
