+++ app/src/main/java/com/websarva/wings/android/reversingvulnapp/viewmodel/vulnalg/VulnAlgViewModel.kt
@@ -8,6 +8,7 @@ import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.Transformations
+import com.websarva.wings.android.reversingvulnapp.viewmodel.SecureSecretKey
 import dagger.hilt.android.lifecycle.HiltViewModel
 import java.io.FileOutputStream
 import java.util.*
@@ -69,69 +70,17 @@ class VulnAlgViewModel @Inject constructor(application: Application): AndroidVie
             }
         }
     }
+    private fun xorDecode(keyPass: String): String{
+        val xorDataByte = String(Base64.decode(getXorData(), Base64.DEFAULT)).toByteArray(Charsets.US_ASCII)
+        var retValue = ""
+
+        for (i in xorDataByte.indices){
+            retValue += String(byteArrayOf((String(Base64.decode(keyPass, Base64.DEFAULT)).toByteArray(Charsets.US_ASCII)[i].xor(xorDataByte[i]))), Charsets.US_ASCII)
+        }
+
+        return retValue
+    }
 
     private external fun getAESData(flag: Int): String
-}
-
-private fun xorDecode(keyPass: String): String{
-    val xorDataByte = String(Base64.decode(getXorData(), Base64.DEFAULT)).toByteArray(Charsets.US_ASCII)
-    var retValue = ""
-
-    for (i in xorDataByte.indices){
-        retValue += String(byteArrayOf((String(Base64.decode(keyPass, Base64.DEFAULT)).toByteArray(Charsets.US_ASCII)[i].xor(xorDataByte[i]))), Charsets.US_ASCII)
-    }
-
-    return retValue
-}
-
-private external fun getXorData(): String
-
-class SecureSecretKey(key2: ByteArray, algorithm: String): SecretKey, Destroyable {
-    private var key: ByteArray?
-    private val algorithm: String
-
-    override fun getAlgorithm(): String {
-        return algorithm
-    }
-
-    override fun getFormat(): String {
-        return "RAW"
-    }
-
-    override fun getEncoded(): ByteArray {
-        if (key == null){
-            throw NullPointerException()
-        }
-        return key!!.clone()
-    }
-
-    override fun destroy() {
-        if (isDestroyed){
-            return
-        }
-        val dummy = "E/AndroidRuntime: FATAL EXCEPTION: main Process: " +
-                "com.websarva.wings.android.androidkeystoresample_kotlin, PID: 6147 " +
-                "java.lang.RuntimeException"
-        val nonSecret: ByteArray = dummy.toByteArray(Charsets.ISO_8859_1)
-
-        for (i in key!!.indices){
-            key!![i] = nonSecret[i % nonSecret.size]
-        }
-        val out= FileOutputStream("/dev/null")
-        out.write(key)
-        out.flush()
-        out.close()
-        key = null
-        System.gc()
-    }
-
-    override fun isDestroyed(): Boolean {
-        return key == null
-    }
-
-    init {
-        this.key = key2.clone()
-        this.algorithm = algorithm
-        Arrays.fill(key2, 0)
-    }
+    private external fun getXorData(): String
 }
